Correctly enumerate audio devices (core + wasapi)
