+++ Cargo.toml
@@ -11,6 +11,7 @@ keywords = ["audio", "sound"]
 
 [dependencies]
 libc = "*"
+lazy_static = "0.1"
 
 [target.i686-pc-windows-gnu.dependencies]
 winapi = "0.2.1"
